debug flag; includes and excludes cmd line arguments added

+++ dot_dependencies.sh
@@ -12,10 +12,16 @@
 # @see: https://maven.apache.org/plugins/maven-dependency-plugin/tree-mojo.html
 
 #set -x
+debug=0
+
 input_files=('pom.xml')
 output_file='dependencies.dot'
 # filter for atifacts
-include='de.icongmbh.*:::'
+includes='de.icongmbh.*:::'
+excludes=''
+
+# page size; include in DOT file as page="..."; default DIN A 4
+page_size='8.3,11.7'
 
 exec_mvn='mvn -B dependency:tree -DoutputType=dot -DappendOutput=true'
 
@@ -42,16 +48,25 @@ exec_cut_console_message='cut -d"]" -f2'
 
 show_help() {
 cat << EOF
-Usage: ${0##*/} [-hsu [-f OUTFILE] [FILE]...
+Usage: ${0##*/} [-dhsu [-e EXCLUDES] [-i INCLUDES] [-o OUTFILE] [-p PAGE_SIZE] [FILE]...
 Create a DOT file based on Maven dependencies (as a 'subgraph') provided by FILE.
 
 With no FILE the default '$input_files' will be used.
     
-    -h          display this help and exit
-    -o OUTFILE  write the result to OUTFILE instead of '$output_file' (default).
-    -s          ONLY SNAPSHOT versions
-    -u          Forces a check for updated releases and snapshots on remote
-                Maven repositories
+    -d           debug option
+    -e EXCLUDES  A comma-separated list of artifacts to filter from the serialized dependency tree, or null (default) not
+                 to filter any artifacts from the dependency tree. An empty pattern segment is treated as an implicit wildcard.
+    -h           display this help and exit
+    -i INCLUDES  A comma-separated list of artifacts to filter the serialized dependency tree by, or null not
+                 to filter the dependency tree. An empty pattern segment is treated as an implicit wildcard.
+                 '$includes' (default)
+                 See: https://maven.apache.org/plugins/maven-dependency-plugin/tree-mojo.html#includes for more information
+    -o OUTFILE   Write the result to OUTFILE instead of '$output_file' (default).
+    -p PAGE_SIZE The page size in inch.
+                 See: http://www.graphviz.org/content/attrs#dpage for more information
+    -s           ONLY SNAPSHOT versions
+    -u           Forces a check for updated releases and snapshots on remote
+                 Maven repositories
     
 Example: ${0##*/} \`find . -mindepth 2 -iname pom.xml | grep -v "target"\`
 EOF
@@ -61,14 +76,22 @@ EOF
 # process command line arguments
 # @see: http://stackoverflow.com/questions/192249/how-do-i-parse-command-line-arguments-in-bash#192266
 OPTIND=1         # Reset in case getopts has been used previously in the shell.
-while getopts "h?vo:" opt;
+while getopts "deh?iopsu:" opt;
 do
     case "$opt" in
+    d)  debug=$((debug + 1))
+        ;;
+    e)  exec_mvn="$exec_mvn -Dexcludes=\"$OPTARG\""
+        ;;
     h|\?)
         show_help
         exit 0
         ;;
+    i)  includes="$OPTARG"
+        ;;
     o)  output_file=$OPTARG
+        ;;
+    p)  page_size="$OPTARG"
         ;;
     s)  include="$include*-SNAPSHOT"
         ;;
@@ -86,6 +109,9 @@ if [ "${#}" -gt "1" ];
 then
   input_files=("$@")
 fi
+
+# add Maven debug option if 'd' command line arg iwas more than once
+[[ $debug -gt 1 ]] && exec_mvn="$exec_mvn -X"
 ### CMD ARGS
 
 ### DEPENDENCIES
@@ -100,8 +126,8 @@ do
   # -DoutputFile and -Doutput seems not work in this special behaviour :-(
   #    mvn_cmd="$exec_mvn -DoutputFile=$temp_output_file -Doutput=$temp_output_file -Dincludes=\"$include\" -f \"$pom_file\" "
   # use the console output instead
-  mvn_cmd="$exec_mvn -Dincludes=\"$include\" -f\"$pom_file\" 2>&1 | $exec_grep_filter_console_message | $exec_cut_console_message"
-#  echo "$mvn_cmd"
+  mvn_cmd="$exec_mvn -Dincludes=\"$includes\" -f\"$pom_file\" 2>&1 | $exec_grep_filter_console_message | $exec_cut_console_message"
+  [[ $debug -gt 0 ]] && echo "$mvn_cmd"
   eval $mvn_cmd >> $temp_output_file
   [[ $? -gt 0 ]] && exit $?; # check the return value
   counter=$((counter + 1))
@@ -117,14 +143,14 @@ fi
 ### CLEAN UP
 echo "create: $output_file"
 echo -e 'digraph G { \n ' > $output_file
-echo -e '    graph [fontsize=8 fontname="Courier" compound=true];\n    node [shape=record fontsize=8 fontname="Courier"];\n    rankdir="LR";\n    page="8.3,11.7";\n ' >> $output_file
+echo -e "    graph [fontsize=8 fontname=\"Courier\" compound=true];\n    node [shape=record fontsize=8 fontname=\"Courier\"];\n    rankdir=\"LR\";\n    page=\"$page_size\";\n " >> $output_file
 # cleanup and normalize DOT content
 cmd_cleanup="$exec_sed_rename_graph $temp_output_file | $exec_sed_normalize_artifacts | $exec_awk_duplicate_lines | $exec_sed_duplicate_braces_line"
-#echo "$cmd_cleanup"
+[[ $debug -gt 0 ]] && echo "$cmd_cleanup"
 eval $cmd_cleanup >> $output_file
 echo '}' >> $output_file
 
-# clean up temp. work file
+# clean up temp. work file if debug level is lower than '2'
 # @see: http://www.linuxjournal.com/content/use-bash-trap-statement-cleanup-temporary-files
-trap "rm -f $temp_output_file" EXIT
+[[ $debug -lt 2 ]] && trap "rm -f $temp_output_file" EXIT
